backend/requirements.txt
anyio
blinker==1.7.0
Bottleneck
certifi
charset-normalizer==3.3.2
click
distro
Flask
Flask-Cors
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
h11
httpcore
httpx
idna
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2
MarkupSafe==2.1.5
mock==5.1.0
numexpr
numpy==1.26.4
openai
packaging==23.2
pandas==2.1.4
pandas-stubs
pluggy==1.4.0
pydantic
PyJWT==2.8.0
pytest==8.0.2
python-dateutil
python-dotenv==1.0.1
pytz
requests==2.31.0
six
sniffio
tqdm
types-pytz
typing_extensions
tzdata
urllib3==2.2.1
Werkzeug